requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
sqlalchemy[asyncio]==2.0.43
asyncpg==0.30.0
psycopg[binary]==3.2.9
celery==5.5.3
flower==2.0.1
elasticsearch[async]==8.17.2
redis==6.4.0
python-dotenv==1.1.1
orjson==3.11.1
pydantic-settings==2.10.1
